Add filedata-tool.py for extracting and creating filedata.bin files.
When extracting a filedata.bin file, all files referenced in the file table located in the game's ELF will be extracted in additon to "hidden" data missing from the file table. A CSV file will be created with IDs, offsets, and lengths found in the game's ELF, hidden file offsets and lengths, the exported filename, and a guessed description of the file contents to aid in modification.

When creating a filedata.bin file, the CSV created by the extraction mode is used to assemble a new filedata.bin file. The file table in the game's ELF is updated to reflect any changes in filesizes.

Tips:
* Don't modify the ids, offsets, or lengths in the CSV file created by the extraction mode. The filenames can be changed if desired.
* New entries cannot be added to the filetable, although this wouldn't be useful anyway. Instead, existing entries must be modified.